StudyWave AI/v2/requirements.txt
# git+https://github.com/openai/whisper.git # need to have git installed for this to work
openai-whisper
moviepy
transformers
fastapi
uvicorn
openai==0.28
python-dotenv